+++ yfinance/base.py
@@ -181,12 +181,6 @@ class TickerBase:
         if params["interval"] == "30m":
             params["interval"] = "15m"
 
-        # setup proxy in requests format
-        if proxy is not None:
-            if isinstance(proxy, dict) and "https" in proxy:
-                proxy = proxy["https"]
-            proxy = {"https": proxy}
-
         # if the ticker is MUTUALFUND or ETF, then get capitalGains events
         params["events"] = "div,splits,capitalGains"
 
@@ -212,6 +206,7 @@ class TickerBase:
             data = get_fn(
                 url=url,
                 params=params,
+                proxy=proxy,
                 timeout=timeout
             )
             if "Will be right back" in data.text or data is None:
@@ -1684,9 +1679,9 @@ class TickerBase:
         return data
 
     @property
-    def fast_info(self):
+    def get_fast_info(self, proxy=None):
         if self._fast_info is None:
-            self._fast_info = FastInfo(self)
+            self._fast_info = FastInfo(self, proxy=proxy)
         return self._fast_info
 
     @property
@@ -1905,7 +1900,7 @@ class TickerBase:
         logger = utils.get_yf_logger()
 
         # Process dates
-        tz = self._get_ticker_tz(proxy=None, timeout=10)
+        tz = self._get_ticker_tz(proxy=proxy, timeout=10)
         dt_now = pd.Timestamp.utcnow().tz_convert(tz)
         if start is not None:
             start_ts = utils._parse_user_dt(start, tz)
@@ -1929,8 +1924,8 @@ class TickerBase:
         ts_url_base = f"https://query2.finance.yahoo.com/ws/fundamentals-timeseries/v1/finance/timeseries/{self.ticker}?symbol={self.ticker}"
         shares_url = f"{ts_url_base}&period1={int(start.timestamp())}&period2={int(end.timestamp())}"
         try:
-            json_str = self._data.cache_get(shares_url).text
-            json_data = _json.loads(json_str)
+            json_data = self._data.cache_get(url=shares_url, proxy=proxy)
+            json_data = json_data.json()
         except (_json.JSONDecodeError, requests.exceptions.RequestException):
             logger.error(f"{self.ticker}: Yahoo web request for share count failed")
             return None
@@ -2099,10 +2094,10 @@ class TickerBase:
 
         return dates
 
-    def get_history_metadata(self) -> dict:
+    def get_history_metadata(self, proxy=None) -> dict:
         if self._history_metadata is None:
             # Request intraday data, because then Yahoo returns exchange schedule.
-            self.history(period="1wk", interval="1h", prepost=True)
+            self.history(period="1wk", interval="1h", prepost=True, proxy=proxy)
 
         if self._history_metadata_formatted is False:
             self._history_metadata = utils.format_history_metadata(self._history_metadata)