+++ yfinance/scrapers/holders.py
@@ -35,7 +35,7 @@ class Holders:
     def _scrape(self, proxy):
         ticker_url = f"{self._SCRAPE_URL_}/{self._data.ticker}"
         try:
-            resp = self._data.cache_get(ticker_url + '/holders', proxy)
+            resp = self._data.cache_get(ticker_url + '/holders', proxy=proxy)
             holders = pd.read_html(resp.text)
         except Exception:
             holders = []